fix(random_delay): Random delay cannot be disabled
Random delay was applied either when 'production_use' is set to true or
'debug' set to false.
The only way to disable it was to set 'production_use' to false and 'debug'
to true.
However, once we enable debug, the search engine can't be used anymore since
it doesn't return any results.