fix(fonts): Serve fonts from a local directory
Currently there is one font in use, and it was served from
http://fonts.gstatic.com/
That adds an extra latency to loading the Websurfx web page.
Also, if the font gets removed from the remote location, it will affect
Websurfx.
The font will now be served from context path '/static/fonts/rubik.woff2'.