fix(page-fetch) Always fetch next page asynchronously
Fetch next page asynchronously.
A delay of 1 sec will be applied when fetching page 1.
This is to prevent page 0 and page 1 to be fetched at the same time or page 1
to be fetched before page 0.
Currently page 0 and page 1 are fetched asynchronously, but we wait both
to complete.
That leads to 2 requests sent to the search engine almost at the same time.

Example:

time: 1714739362, Page: 1, URL: https://duckduckgo.com/html/?q=ECB&s=30&dc=31&v=1&o=json&api=/d.js
time: 1714739363, Page: 0, URL: https://html.duckduckgo.com/html/?q=ECB&s=&dc=&v=1&o=json&api=/d.js

Some search engines (E.g. Qwant), may recognize Websurfix as a bot and redirect
us to a challenge to prove otherwise.

Also, since we are waiting page 1 to be loaded before we display the results
from page 0, we are increasing the load time for page 0.
Similar is the situation when we try to load the results from current_page + 1.
Although they are already in the cache, we are waiting the results for current_page + 2 to
be loaded, and then we display the results for current_page + 1.
Hense, affecting the user experience.