fix(fake_useragent) Load the fake user agents during startup
Currently the fake user agents are loaded after we make the first
search request since the start of Websurfx.
That leads to a longer time to load the results.
Usually more than 2 sec.
Hense, that leads to a bad user experience.
It is perceived, that for a good user experience, a page should
load in less than 2 sec.
The user agents are loaded asynchronously, in order not to prolong
the startup time of Websurfx.