fix(header): Add Accept-Language header to the upstream requests
Get the value of Accept-Language header from the user and add it
to the header_map for the requests sent to the upstream search
engines.
Currently the requests sent to the search engines don't include
language information.
However, some of the randomly used fake User-Agent strings may
include language information which will be interpreted by the
upstream search engine.
In such cases we may receive results in a language
different from the one used in our requests.
That would make the sorting of the results inefficient.
Also, some search engines (Eg. Qwant) are offered in limited list
of languages and if we don't provide any language, or a language
which is not in that list, we will receive an error that our
language is not supported.